+++ hw3/hw3-1/forEach.js
@@ -1,0 +1,7 @@
+function forEach(arr, callback) {
+    for (var i = 0; i < arr.length; i++) {
+        callback(arr[i], i, arr);
+    }
+}
+
+module.exports = forEach;
