+++ hw3/hw3-1/map.js
@@ -1,0 +1,11 @@
+const push = require ("./push");
+
+function map(arr, callback) {
+    var newArr = [];
+    for (var i = 0; i < arr.length; i++) {
+        push(newArr, callback(arr[i], i, arr));
+    }
+    return newArr;
+}
+
+module.exports = map;
