+++ hw3/hw3-1/push.js
@@ -1,0 +1,7 @@
+function push(arr) {
+    for (var i = 0; i < arguments.length - 1; i++) {
+        arr[arr.length] = arguments[i + 1];
+    }
+}
+
+module.exports = push;
